Fixed bug in shelvedb introduced by previous commit. Started adding metadata grabbing functions.
